Updated python script to use new "extract directory" mode.
Additionally fixed oversight with audio files and bug in OTRGui.
